Support nix-elisp-bundle to set package-directory-list/load-path
The Nix package manager has a nice way to bundle elisp packages and their
dependencies for a particular version or build of Emacs. The way it normally
works is to wrap the executable with a script that sets EMACSLOADPATH and other
variables. But it can also be made to support profiles with different package
sets when using chemacs.